refactor: encapsulated CLI into logical hierarchy breaking methods/functions into their own subsections
